Annuler Transaction faire une Transaction

+++ src/main/java/com/backendwave/data/repositories/NumeroFavoriRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 public interface NumeroFavoriRepository extends BaseRepository<NumeroFavori> {
 
     // Trouver tous les numéros favoris d'un client
-    List<NumeroFavori> findByClientId(Long clientId);
+    List<NumeroFavori> findByClientId(int clientId);
 
     // Trouver un numéro favori spécifique d'un client
     Optional<NumeroFavori> findByClientIdAndNumeroTelephone(Long clientId, String numeroTelephone);
